Switch AppI18n to I18n & remove network_lookup if remote translations disabled
