+++ packages/rx_bloc_cli/mason_templates/bricks/feature_deeplink/__brick__/lib/l10n/{{#enable_remote_translations}}network_lookup{{/enable_remote_translations}}/enter_message_lookup.dart
@@ -1,0 +1,33 @@
+{{> licence.dart}}
+
+import '../../assets.dart';
+import '../{{project_name}}_app_i18n.dart';
+import 'util.dart';
+
+class AppI18nEnterMessageLookup extends I18nFeatureEnterMessageLookup {
+  @override
+  String getString(String key, [Map<String, String>? placeholders]) {
+    return getFromNetwork('enter_message__', key, placeholders) ??
+        getFromBundle(AppI18n.locale?.languageCode)
+            .getString(key, placeholders)!;
+  }
+
+  I18nFeatureEnterMessage? bundledI18n;
+
+  I18nFeatureEnterMessage getFromBundle(String? languageCode) {
+    if (bundledI18n == null) {
+      switch (I18n.currentLocale?.languageCode) {
+        case 'bg':
+          bundledI18n =
+              I18nFeatureEnterMessage(I18nFeatureEnterMessageLookup_bg());
+          break;
+        case 'en':
+        default:
+          bundledI18n =
+              I18nFeatureEnterMessage(I18nFeatureEnterMessageLookup_en());
+          break;
+      }
+    }
+    return bundledI18n!;
+  }
+}
