+++ packages/rx_bloc_cli/mason_templates/bricks/feature_deeplink/brick.yaml
@@ -6,4 +6,3 @@ version: 0.1.0+1
 environment:
   mason: ">=0.1.0-dev.47 <0.1.0"
 vars:
-  - enable_remote_translation
