+++ packages/rx_bloc_cli/mason_templates/bricks/feature_otp/brick.yaml
@@ -8,4 +8,3 @@ environment:
   mason: ">=0.1.0-dev.49 <0.1.0"
 
 vars:
-  - enable_remote_translation
