+++ packages/rx_bloc_cli/mason_templates/bricks/feature_showcase/brick.yaml
@@ -7,4 +7,3 @@ environment:
   mason: ">=0.1.0-dev.60 <0.1.0"
 
 vars:
-  - enable_remote_translation
