+++ packages/rx_bloc_cli/mason_templates/bricks/feature_widget_toolkit/brick.yaml
@@ -25,4 +25,3 @@ environment:
 #  * an optional prompt phrase used when asking for the variable
 #  * a list of values (enums only)
 vars:
-  - enable_remote_translation
