+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/app/{{project_name}}.dart
@@ -11,8 +11,8 @@ import 'package:local_session_timeout/local_session_timeout.dart';{{/enable_pin_
 import 'package:provider/provider.dart'; {{#enable_change_language}}
 import 'package:widget_toolkit/language_picker.dart'; {{/enable_change_language}}
 
-import '../../assets.dart';
-import '../../l10n/{{project_name}}_app_i18n.dart';{{#analytics}}
+import '../../assets.dart';{{#enable_remote_translations}}
+import '../../l10n/{{project_name}}_app_i18n.dart';{{/enable_remote_translations}}{{#analytics}}
 import '../../lib_analytics/blocs/analytics_bloc.dart';{{/analytics}}{{#has_authentication}}{{#enable_pin_code}}
 import '../../lib_auth/blocs/user_account_bloc.dart';{{/enable_pin_code}}
 import '../../lib_auth/data_sources/remote/interceptors/auth_interceptor.dart';{{/has_authentication}} {{#enable_change_language}}
@@ -256,8 +256,9 @@ Widget _buildMaterialApp(BuildContext context) => MaterialApp.router(
        title: '{{#titleCase}}{{project_name}}{{/titleCase}}',
        theme: {{project_name.pascalCase()}}Theme.buildTheme(DesignSystem.light()),
        darkTheme: {{project_name.pascalCase()}}Theme.buildTheme(DesignSystem.dark()),
-       localizationsDelegates: const [
-         AppI18n.delegate,
+       localizationsDelegates: const [ {{#enable_remote_translations}}
+         AppI18n.delegate,{{/enable_remote_translations}}{{^enable_remote_translations}}
+         I18n.delegate,{{/enable_remote_translations}}
          ...GlobalMaterialLocalizations.delegates,
        ],
        supportedLocales: I18n.supportedLocales,
