+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/l10n.dart
@@ -1,9 +1,11 @@
 {{> licence.dart }}
 
 import 'package:flutter/widgets.dart';
+{{#enable_remote_translations}}
+import '{{project_name}}_app_i18n.dart';{{/enable_remote_translations}}{{^enable_remote_translations}}
+import '../assets.dart';{{/enable_remote_translations}}
 
-import '{{project_name}}_app_i18n.dart';
-
-extension AppLocalizationsX on BuildContext {
-  AppI18n get l10n => AppI18n.of(this);
+extension AppLocalizationsX on BuildContext { {{#enable_remote_translations}}
+  AppI18n get l10n => AppI18n.of(this);{{/enable_remote_translations}}{{^enable_remote_translations}}
+  I18n get l10n => I18n.of(this);{{/enable_remote_translations}}
 }
