--- packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/network_lookup/{{#has_authentication}}login_lookup.dart{{/has_authentication}}
@@ -1,33 +1,0 @@
-{{> licence.dart}}
-
-import '../../assets.dart';
-import '../{{project_name}}_app_i18n.dart';{{#enable_remote_translations}}
-import 'util.dart';{{/enable_remote_translations}}
-
-class AppI18nLoginLookup extends I18nFeatureLoginLookup {
-  @override
-  String getString(String key, [Map<String, String>? placeholders]) { {{#enable_remote_translations}}
-    return getFromNetwork('login__', key, placeholders) ??
-        getFromBundle(AppI18n.locale?.languageCode)
-            .getString(key, placeholders)!;{{/enable_remote_translations}}{{^enable_remote_translations}}
-    return getFromBundle(AppI18n.locale?.languageCode)
-        .getString(key, placeholders)!;{{/enable_remote_translations}}
-  }
-
-  I18nFeatureLogin? bundledI18n;
-
-  I18nFeatureLogin getFromBundle(String? languageCode) {
-    if (bundledI18n == null) {
-      switch (I18n.currentLocale?.languageCode) {
-        case 'bg':
-          bundledI18n = I18nFeatureLogin(I18nFeatureLoginLookup_bg());
-          break;
-        case 'en':
-        default:
-          bundledI18n = I18nFeatureLogin(I18nFeatureLoginLookup_en());
-          break;
-      }
-    }
-    return bundledI18n!;
-  }
-}
